+++ test/default_manifold.jl
@@ -887,4 +887,38 @@ Base.size(x::MatrixVectorTransport) = (size(x.m, 2),)
         @test isapprox(vee(MC, p, [1 + 2im, 3 + 4im, 5 + 6im]), [1, 3, 5, 2, 4, 6])
         @test isapprox(hat(MC, p, [1, 3, 5, 2, 4, 6]), [1 + 2im, 3 + 4im, 5 + 6im])
     end
+
+    ManifoldsBase.default_approximation_method(
+        ::ManifoldsBase.DefaultManifold,
+        ::typeof(exp),
+    ) = GradientDescentEstimation()
+    @testset "Estimation Method defaults" begin
+        M = ManifoldsBase.DefaultManifold(3)
+        # Point generic type fallback
+        @test default_approximation_method(M, exp, Float64) ==
+              default_approximation_method(M, exp)
+        # Retraction
+        @test default_approximation_method(M, retract) == default_retraction_method(M)
+        @test default_approximation_method(M, retract, DefaultPoint) ==
+              default_retraction_method(M)
+        # Inverse Retraction
+        @test default_approximation_method(M, inverse_retract) ==
+              default_inverse_retraction_method(M)
+        @test default_approximation_method(M, inverse_retract, DefaultPoint) ==
+              default_inverse_retraction_method(M)
+        # Vector Transsports – all 3: to
+        @test default_approximation_method(M, vector_transport_to) ==
+              default_vector_transport_method(M)
+        @test default_approximation_method(M, vector_transport_to, DefaultPoint) ==
+              default_vector_transport_method(M)
+        # along
+        @test default_approximation_method(M, vector_transport_along) ==
+              default_vector_transport_method(M)
+        @test default_approximation_method(M, vector_transport_along, DefaultPoint) ==
+              default_vector_transport_method(M)
+        @test default_approximation_method(M, vector_transport_direction) ==
+              default_vector_transport_method(M)
+        @test default_approximation_method(M, vector_transport_direction, DefaultPoint) ==
+              default_vector_transport_method(M)
+    end
 end